Add files via upload
The schematic and software for the tube furnace. 

Software:
The software for now only contains the basic setup including encoder debouncing, LED test, and I2C communication with the LCD. The encoder debouncing and LED test were done through timer interrupts.

Schematic: 
Mostly self-explanatory. The data logger fits on top of the Arduino and the outlets on the data logger are soldered to the corresponding pins on the Arduino. There is no schematic for the plug and fuse and relay connection yet.