Tube Furnace Schematic and Software
Schematic:
Pretty self-explanatory. The data logger is supposed to fit onto the Arduino. The outlets soldered onto the data logger are also soldered to the corresponding pins on the Arduino. No schematic for the plug and fuse and relay connection yet.

Code:
Basic setup code. Includes: LED test, encoder debouncing, and I2C communication with the LCD and encoder. The LED test and encoder debounce were done through timer interrupts.